requirements.txt
python-dotenv
discord.py
psycopg2-binary
sqlalchemy
gspread
oauth2client